+++ 임재욱/book/src/main/java/book/Application.java
@@ -1,4 +1,4 @@
-package umc.book;
+package book;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
