+++ 임재욱/book/src/main/java/book/domain/BookCategory.java
@@ -1,11 +1,8 @@
-package umc.book.domain;
+package book.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.book.domain.base.BaseEntity;
-
-import java.time.LocalDateTime;
-import java.util.List;
+import book.domain.base.BaseEntity;
 
 
 @Getter
