+++ 임재욱/book/src/main/java/book/domain/Hashtag.java
@@ -1,8 +1,8 @@
-package umc.book.domain;
+package book.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.book.domain.base.BaseEntity;
+import book.domain.base.BaseEntity;
 
 
 @Getter
