+++ 임재욱/book/src/main/java/book/domain/PersonalAlarm.java
@@ -1,11 +1,9 @@
-package umc.book.domain;
+package book.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.book.domain.base.BaseEntity;
-import umc.book.domain.enums.AlarmKind;
-
-import java.time.LocalDateTime;
+import book.domain.base.BaseEntity;
+import book.domain.enums.AlarmKind;
 
 @Getter
 @Entity
