+++ 임재욱/book/src/main/java/book/domain/Rent.java
@@ -1,8 +1,8 @@
-package umc.book.domain;
+package book.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.book.domain.base.BaseEntity;
+import book.domain.base.BaseEntity;
 
 import java.time.LocalDateTime;
