+++ 임재욱/book/src/main/java/book/domain/base/BaseEntity.java
@@ -1,4 +1,4 @@
-package umc.book.domain.base;
+package book.domain.base;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
