+++ 임재욱/book/src/main/java/book/domain/enums/Gender.java
@@ -1,3 +1,3 @@
-package umc.book.domain.enums;
+package book.domain.enums;
 
 public enum Gender { MALE, FEMALE, OTHER }
